requirements.txt
requests
dotenv
langchain-openai
azure-identity
beautifulsoup4
selenium
streamlit